requirements.txt
matplotlib>=2.0.2
numpy>=1.12.1
scipy>=0.19.0
Pillow>=3.1.2
